Add player guess function
This commit when implemented will add the player `#player_guess` method to the human class.

This method will ask the user to input their code guess. User inputs are not case sensitive and if user inputs invalid key selection, user will be prompted for a valid selection until user provides a valid input.