Add partial and match functions
When implemented this commit will add the partial match and full match functions. The purpose of these functions is detailed below.

`#partial_match`
Check the guess of the code breaker for keys that are the correct color but not in the correct position. This function return how many partial_matches the code breaker got in the form of an int.

`#matches`
This function checks how many keys in teh code_breakers guess are the correct color and in the correct position. This function returns the number of matches as an int.